+++ CHANGELOG.md
@@ -3773,3 +3773,4 @@
 - [2026-08-20 15:51] test(theme): add unit tests for theme business logic
 - [2026-08-20 16:03] feat(test): add async test stream provider
 - [2026-08-20 16:26] test(cache): implement widget integration tests for cache
+- [2026-08-20 16:30] fix(forms): resolve race condition during forms initialization